+++ spring-nanotrader-data/src/main/java/org/springframework/nanotrader/data/repository/HoldingAggregateRepository.java
@@ -25,6 +25,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface HoldingAggregateRepository {
 
-	abstract HoldingSummary findHoldingAggregated(Integer accountId);
+	abstract HoldingSummary findHoldingAggregated(Long accountId);
 
 }
