+++ spring-nanotrader-data/src/main/java/org/springframework/nanotrader/data/repository/PortfolioSummaryRepository.java
@@ -25,6 +25,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface PortfolioSummaryRepository {
 
-	abstract PortfolioSummary findPortfolioSummary(Integer accountId);
+	abstract PortfolioSummary findPortfolioSummary(Long accountId);
 
 }
